requirements.txt
Flask
Flask-SQLAlchemy
Flask-Mail
Flask-WTF
WTForms
email-validator
python-dotenv
Flask-Login
Flask-Migrate
Flask-Bcrypt
Flask-LDAP3-Login
gunicorn

